requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.9.0
openai==1.76.0
pillow==11.2.1
playwright==1.51.0
pyairtable==3.1.1
pydantic==2.11.3
pydantic_core==2.33.1
pyee==12.1.1
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
